requirements.txt
streamlit>=1.37
pandas>=2.0
numpy>=1.26
yfinance>=0.2.50
python-dotenv>=1.0
tenacity>=8.2
httpx>=0.27
pydantic>=2.8
sqlmodel>=0.0.22

pytest>=8.0
pytest-mock>=3.14
